requirements.txt
Flask==1.1.1
numpy==1.19.4
matplotlib==3.2.2
torch==1.7.0
torchvision==0.8.1
Pillow==7.1.0
flask_cors==3.0.10
opencv-python==4.5.1.48
keras==2.4.3
tensorflow==2.4.1
wandb==0.10.12
tqdm==4.55.1
helper==2.4.2
cycler==0.11.0
pandas==1.1.4
scikit_learn==0.23.2
seaborn==0.11.0
MulticoreTSNE==0.1
